menina-do-rh/requirements.txt
faiss-cpu
sentence-transformers
python-dotenv
requests
